requirements_dev.txt
-r requirements.txt

ipython==7.28.0
flake8==4.0.1
isort==5.9.3
pytest==6.2.5
pytest-asyncio==0.16.0
pytest-cov==3.0.0
pytest-aiohttp==0.3.0
